broke out classes from workout_screen
